Updated Risk Management Documents for Sprint 2
The log and management plan for risks have been updated. @adrientremblay and I added an Updated section for each risk. Some risks are marked as Happening and have been tended to for most of the sprint. The effect of the response strategies will be collected in between sprint 02 and sprint 03.